fixed bugs, got final version to work
App has a form for uploading files to a database. With each upload, a timestamp, and a random password is generated in to the database. The generated password is stored in the database as a crypted hash of the randomly generated password. When retrieving files from the database, the id of the wanted file needs to be known, and the password for acquiring the specific file.  When the right information is given, the user is shown a response body with the file on the page.